fix: 페이징과 같이 쓸수 있도록 리크루트 목록 조회시 next의 타입 Integer로 변경

+++ src/main/java/com/ssafy/ssafsound/domain/recruit/dto/GetRecruitsReqDto.java
@@ -10,7 +10,7 @@ import java.util.List;
 @Getter
 @Builder
 public class GetRecruitsReqDto {
-    private Long next;
+    private Integer next;
 
     @NotBlank
     private String category;
